TournamentFinder/requirements/tournamentfinder.txt
pip==24.0

django==5.0.2
django-localflavor==4.0
django-phonenumber-field[phonenumberslite]==7.3.0
pillow==10.2.0